Rename GetPrefilledDraftDataPresenter to GetDraftDetailsPresenter
This GetPrefilledDraftDataPresenter was confusingly named as it was
really the presenter responsible for rendering draft details.  This
issue was addressed by renaming the presenter in question and the
associated modules and tests.
